test: cover NetworkStack/IP/ARP via a TAP I/O seam, add UDP outbound ARP, and tidy
Introduce a PacketChannel seam so NetworkStack can be driven in tests with a
fake frame transport instead of a real TAP device, then add the previously
missing integration coverage for the core wiring layer.

- PacketChannel interface + FakePacketChannel double; NetworkStack takes an
  injectable channel and the existing string constructor delegates through it,
  so no call sites change
- New tests: test_network_stack (ARP reply, full TCP handshake + accept, RST,
  ICMP echo, UDP port-unreachable, UDP delivery, and the UDP/connect
  resolve-then-flush paths), test_ip, test_arp
- UDP send_to() to an unresolved peer now resolves it via the same bounded ARP
  request/retry/queue machinery connect() uses, instead of dropping the datagram
- Cleanup: unify the IP flag bits into one field (fixing the constructor that
  ignored its version and reserved-flag arguments), add Makefile all/test
  targets, and silence two benign sign-compare warnings for a clean -Wall build

Verified 73/73 tests pass and the build is warning-free under gcc 13.

+++ include/utils.h
@@ -46,10 +46,10 @@ T bytes_to_int(const Bytes& bytes)
 {
     T result = 0;
     size_t byte_count = sizeof result;
-    for (int i = 0; i < byte_count; i++)
+    for (size_t i = 0; i < byte_count; i++)
     {
         result *= (1 << 8);
-        result += bytes[i]; 
+        result += bytes[i];
     }
     return result; 
 }
